some ui improvements, mediaHandling, fileWidgets
