+++ packages/repco-frontend/package.json
@@ -19,6 +19,7 @@
     "@radix-ui/react-dropdown-menu": "^2.0.2",
     "@radix-ui/react-icons": "^1.1.1",
     "@radix-ui/react-radio-group": "^1.1.1",
+    "@radix-ui/react-toast": "^1.1.2",
     "@radix-ui/react-tooltip": "^1.0.3",
     "@remix-run/node": "^1.4.3",
     "@remix-run/react": "^1.4.3",
